requirements.txt
opencv-contrib-python
matplotlib
numpy
sympy
screeninfo
python-dotenv
